Add new, "smart" DID spoofer
A DID spoofer that should always work on the latest version. It scans the game for your current DID then saves the pointers. The non-static pointers are used to spoof your DID onwards, even after restarting the game. If you re-open the app, it will re-use the saved pointers if the game's process ID is still the same.